feat: lazy-load supabase client
