+++ src/components/ProtectedRoute.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { Navigate } from 'react-router-dom'
 import { useAuth } from '@/contexts/AuthContext'
 import { LoadingSpinner } from '@/components/ui/LoadingSpinner'
-import { supabase } from '@/lib/supabase'
+import { getSupabaseClient } from '@/lib/supabase'
 
 interface ProtectedRouteProps {
   children: React.ReactNode
@@ -13,22 +13,34 @@ export function ProtectedRoute({ children }: ProtectedRouteProps) {
   const [sessionChecked, setSessionChecked] = useState(false)
 
   useEffect(() => {
-    // Double-check session on mount
+    let active = true
+
     const checkSession = async () => {
       try {
+        const supabase = await getSupabaseClient()
         const { data: { session } } = await supabase.auth.getSession()
+        if (!active) return
+
         if (session) {
           console.log('Session validated in ProtectedRoute')
         }
       } catch (error) {
-        console.error('Session check failed:', error)
+        if (active) {
+          console.error('Session check failed:', error)
+        }
       } finally {
-        setSessionChecked(true)
+        if (active) {
+          setSessionChecked(true)
+        }
       }
     }
-    
+
     if (!loading) {
       checkSession()
+    }
+
+    return () => {
+      active = false
     }
   }, [loading])
